requirements.txt
pyttsx3==2.90
SpeechRecognition==3.8.1
wikipedia==1.4.0
openai==0.10.5
python-dotenv==0.19.1
